refactor scan generation code
