+++ data_processing/radiology/refined_table/__init__.py
@@ -1,0 +1,5 @@
+'''
+Created on November 19, 2020
+
+@author: rosed2@mskcc.org
+'''
