+++ tests/data_processing/testdata/data/test-project/tables/dicom/_delta_log/00000000000000000000.json
@@ -1,0 +1,4 @@
+{"commitInfo":{"timestamp":1605794939162,"operation":"WRITE","operationParameters":{"mode":"Overwrite","partitionBy":"[]"},"isBlindAppend":false,"operationMetrics":{"numFiles":"1","numOutputBytes":"536669","numOutputRows":"1"}}}
+{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
+{"metaData":{"id":"cf37ab42-bf0b-42a4-84a1-aad6f6ce4232","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"path\",\"type\":\"string\",\"nullable\":true,\"metadata\":{}},{\"name\":\"modificationTime\",\"type\":\"timestamp\",\"nullable\":true,\"metadata\":{}},{\"name\":\"length\",\"type\":\"long\",\"nullable\":true,\"metadata\":{}},{\"name\":\"content\",\"type\":\"binary\",\"nullable\":true,\"metadata\":{}},{\"name\":\"dicom_record_uuid\",\"type\":\"string\",\"nullable\":true,\"metadata\":{}},{\"name\":\"metadata\",\"type\":{\"type\":\"map\",\"keyType\":\"string\",\"valueType\":\"string\",\"valueContainsNull\":true},\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1605794934821}}
+{"add":{"path":"part-00000-b5adc7f3-1d48-47f9-816f-aaf0ef90cb93-c000.parquet","partitionValues":{},"size":536669,"modificationTime":1605794937000,"dataChange":true}}
